Factor in object color in diffuse component
From Khadem: Diffuse coefficient is dependent on surface color, that
maybe a reason to use intersectionPointColor in diffuse component. For
same reason, ambient color is `ambient_coeff*intersectionPointColor` in
offline spec. But in offline spec, intersectionPointColor is used in
specular component too, which shouldn't be I believe (as specular
component is independent of surface color, if I recall correctly).