Fix Floor object
The parameters floor_size (size) and tile_size (len) weren't used
properly. I used floor_size as number of tiles in x- or y-direction and
tile_size as the width of a tile.

Now, size parameter and len parameter are used as floor size and tile
size so as for better readability.